+++ test/JsonSerializerTests.cpp
@@ -1,3 +1,23 @@
+// Copyright (c) 2022 OpCoSim
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 #include <gtest/gtest.h>
 #include "OpCoSerializer/OpCoSerializer.hpp"
 
@@ -10,29 +30,81 @@ struct TestTypeWithProperties final
     double d = 1.23;
     bool b = true;
     std::vector<double> v{1.2, 3.4};
+    std::string s;
 
     bool operator==(TestTypeWithProperties const& other) const
     {
         return i == other.i && d == other.d && b == other.b && v == other.v;
     }
 
-    static auto constexpr properties() { return std::make_tuple(
-        MakeProperty(&TestTypeWithProperties::i, "integer"),
-        MakeProperty(&TestTypeWithProperties::d, "double"),
-        MakeProperty(&TestTypeWithProperties::b, "boolean"),
-        MakeProperty(&TestTypeWithProperties::v, "vector")
-    );};
+    static auto constexpr SerializerProperties() { 
+        return std::make_tuple(
+            MakeProperty(&TestTypeWithProperties::i, "integer"),
+            MakeProperty(&TestTypeWithProperties::d, "double"),
+            MakeProperty(&TestTypeWithProperties::b, "boolean"),
+            MakeProperty(&TestTypeWithProperties::v, "vector"),
+            MakeProperty(&TestTypeWithProperties::s, "string")
+        );
+    };
+};
+
+struct Nested final
+{
+    int value;
+
+    bool operator==(Nested const& other) const
+    {
+        return value == other.value;
+    }
+
+    static auto constexpr SerializerProperties() { 
+        return std::make_tuple(MakeProperty(&Nested::value, "value"));
+    };
+};
+
+struct WithNested final
+{
+    Nested nested;
+
+    bool operator==(WithNested const& other) const
+    {
+        return nested == other.nested;
+    }
+
+    static auto constexpr SerializerProperties() { 
+        return std::make_tuple(MakeProperty(&WithNested::nested, "nested"));
+    };
+};
+
+enum class TestEnum
+{
+    Value
+};
+
+struct WithEnum final
+{
+    TestEnum enumValue;
+
+    bool operator==(WithEnum const& other) const
+    {
+        return enumValue == other.enumValue;
+    }
+
+    static auto constexpr SerializerProperties() { 
+        return std::make_tuple(MakeProperty(&WithEnum::enumValue, "enum"));
+    };
 };
 
 TEST(JsonSerializer, SerializesExpectedString)
 {
     JsonSerializer serializer{};
-    auto expected = "{\"integer\":5,\"double\":1.5,\"boolean\":true,\"vector\":[1.2]}";
+    auto expected = "{\"integer\":5,\"double\":1.5,\"boolean\":true,\"vector\":[1.2],\"string\":\"test\"}";
     TestTypeWithProperties value = {
         5,
         1.5,
         true,
-        std::vector<double>{1.2}
+        std::vector<double>{1.2},
+        "test"
     };
 
     auto serialized = serializer.Serialize(value);
@@ -43,12 +115,13 @@ TEST(JsonSerializer, SerializesExpectedString)
 TEST(JsonSerializer, DeerializesExpectedValue)
 {
     JsonSerializer serializer{};
-    std::string string{"{\"integer\":7,\"double\":4.2,\"boolean\":false,\"vector\":[1.2,3.4,4.5]}"};
+    std::string string{"{\"integer\":7,\"double\":4.2,\"boolean\":false,\"vector\":[1.2,3.4,4.5],\"string\":\"example\"}"};
     TestTypeWithProperties expected = {
         7,
         4.2,
         false,
-        std::vector<double>{1.2, 3.4, 4.5}
+        std::vector<double>{1.2, 3.4, 4.5},
+        "example"
     };
 
     auto deserialized = serializer.Deserialize<TestTypeWithProperties>(string);
@@ -68,6 +141,32 @@ TEST(JsonSerializer, RoundTripTest)
     
     auto serialized = serializer.Serialize(value);
     auto deserialized = serializer.Deserialize<TestTypeWithProperties>(serialized);
+
+    ASSERT_EQ(value, deserialized);
+}
+
+TEST(JsonSerializer, NestedRoundTripTest)
+{
+    JsonSerializer serializer{};
+    WithNested value = {
+        Nested {
+            4
+        }
+    };
+    
+    auto serialized = serializer.Serialize(value);
+    auto deserialized = serializer.Deserialize<WithNested>(serialized);
+
+    ASSERT_EQ(value, deserialized);
+}
+
+TEST(JsonSerializer, EnumRoundTripTest)
+{
+    JsonSerializer serializer{};
+    WithEnum value = { TestEnum::Value };
+    
+    auto serialized = serializer.Serialize(value);
+    auto deserialized = serializer.Deserialize<WithEnum>(serialized);
 
     ASSERT_EQ(value, deserialized);
 }