+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### ✨ Added
 
+- `std::string` support in JSON serializer.
+- Native `enum class` support in JSON serializer.
+
+### 🙌 Improvements
+
+- Types with OpCoSerializer properties are now automatically detected and serialized to objects.
+- Renames `properties` to `SerializerProperties` for clearer intent on consuming types.
+
+## 🔖 [[v0.1.0]](https://github.com/OpCoSim/OpCoSerializer/releases/tag/v0.1.0 "v0.1.0 Release")
+
+### ✨ Added
+
 - Initial project 🎂
 - JSON serializer implementation.
 - Extendable type serialization.